is_uuid, function added to the uuid-ossp module for checking if a text value is parseable to a uuid.
This overcomes a specific problem with fast data cleansing.
The use-case is for quickly finding uuid / non-uuid values
in a text column in a large table. Postgres can perform a
parallel scan to improve overall query time, but this is
blocked when any function used in the scan is not parallel
safe.

Trying to compose an existing function into a new one results
in a non-parallel safe function (because it contains a
subtransaction in the form of the exception block), e.g.:

create or replace function is_uuid(str text)
    returns bool
    language plpgsql
    parallel safe
as
$$
begin
    perform str::uuid;
    return true;
exception when invalid_text_representation then
    return false;
end;
$$ ;

results in the following error:

ERROR:  cannot start subtransactions during a parallel operation
CONTEXT:  PL/pgSQL function uuid_or_null(text) line 2 during statement
block entry

An alternative approach is to use a regular expression for matching
uuid format, e.g.:

create or replace function is_uuid_re(str text)
    returns bool
    language plpgsql
    parallel safe
as
$$
begin
    return str !~ '[a-f0-9]{8}(-[a-f0-9]{4}){3}-[a-f0-9]{12}';
end;
$$;

TODO benchmark both of these approaches and provide comparisons in
the commit log. But I expect the regex one to be significantly slower
than the C language one.

The approach taken in this patch is simply to duplicate the canonical
uuid parsing done in core uuid.h, strip unnecessary variables, and
simply return true or false if the parsing was successful. This has
the advantages that:
1. it's quicker than regex and
2. anything returning true from is_uuid() can be cast to a uuid with
some confidence.

The downside is code duplication, thoughts welcome for how to prevent
this.